fixing dtos in course still working on it good stil
