+++ backend/src/app/dtos/general.dto.ts
@@ -1,0 +1,5 @@
+export interface UserDTO {
+  email: string,
+  id: string,
+  role: string,
+}
