+++ backend/src/app/dtos/stats.dto.ts
@@ -1,3 +1,5 @@
+import { CourseStats } from "../../domain/types/course-stats.interface";
+
 export interface Period {
   start: Date;
   end: Date;
@@ -12,18 +14,7 @@ export interface Person {
 
 
 
-export interface CourseStats {
-  courseId: string;
-  courseTitle: string;
-  instructorName?: string;
-  enrollmentCount: number;
-  revenue: number;
-  rating: number;
-  reviewCount: number;
-  status?: string; // For instructor-specific stats
-  createdAt?: Date;
-  lastEnrollmentDate?: Date;
-}
+
 
 export interface InstructorStats {
   instructorId: string;
