+++ backend/src/app/usecases/course/interfaces/get-course-stats.usecase.interface.ts
@@ -1,21 +1,5 @@
-export interface ICourseStats {
-  totalCourses: number;
-  activeCourses: number;
-  inactiveCourses: number;
-  pendingCourses: number;
-  approvedCourses: number;
-  declinedCourses: number;
-  publishedCourses: number;
-  draftCourses: number;
-  archivedCourses: number;
-}
-
-export interface IGetCourseStatsInput {
-  userId?: string;
-  includeDeleted?: boolean;
-  isAdmin?: boolean;
-}
+import {  CourseOverallStatsDTO, IGetCourseStatsInput } from "../../../dtos/course.dto";
 
 export interface IGetCourseStatsUseCase {
-  execute(input: IGetCourseStatsInput): Promise<ICourseStats>;
+  execute(input: IGetCourseStatsInput): Promise<CourseOverallStatsDTO>;
 }
